ktest: remove explict conversion to typeof(a)
This undoes the previous change that allowed KEXPECT_EQ(NULL, ...) to
compile.  By explicitly casting bval to typeof(a), b would be silently
narrowed to typeof(a) --- this meant, for example, that
KEXPECT_EQ(1, 0x100000002) would pass, because the upper 32 bits of bval
are discarded.

In this version (back to what it used to be), it's an implicit
conversion from bval to typeof(a).  This still risks a narrowing, and
if a or b are variables (rather than constants), that won't be caught at
compile-time if extra bits are discarded.  For constants, though,
narrowing will be caught.  So the example above will fail to compile.

KEXPECT_NULL and KEXPECT_NOT_NULL are added as conveniences (since the
natural form no longer compiles, since NULL is void* not const void*).

Ideally we could avoid all of these problems, but I can't figure out a
way to do that easily --- the codebase is absolutely full of comparisons
between signed and unsigned values that fail to compile if we use
typeof(b) for bval.