pci: add TODO for broken memory-mapped BARs on x86
pci-legacy should be putting the virtual rather than physical addresses
into the bar.io field.  However, on x86/x64, qemu maps the physical
PCI memory address space into a very high address (starting at
0xfebd0000 on my test), which can't be access from the basic direct
physical-to-virtual kernel map.

On 64-bit systems it's not hard to simply map more of the physical
address space (this is what happens on riscv64).  For x86, though, we'll
need another approach to create a custom mapping (either dynamically, or
figure out how to statically map the top 32MiB of the address space or
something hacky like that).

Converting these addresses to virtual now causes a panic, and no current
devices (other than in-progress NVMe) use their memory mappings, so
leaving the bug as-is for now with a TODO.