nvme: disable NVMe IRQs after scheduling defint
I previously thought that it was an NVMe or qemu property that IRQs
wouldn't continuously fire for the NVME controller, but on x86 the IRQ
continues to fire.

Rather than rework the driver to empty the queue in the IRQ handler, now
we disable the IRQ until after the defint runs.  This is possibly better
anyway, as it keeps the IRQ handler lightweight (at the cost of IRQ
handling latency, but that seems OK).