nvme: fix use-after-free on interrupt transaction
The transaction is stack-allocated in nvme_submit_blocking(), and when
an operation is interrupted (e.g. by a signal), the memory is still
referenced in the transaction in the pending table.

This manifested previously about 5-10% of the time in the VFS O_CLOEXEC
test due to the SIGKILL.  There is now a dedicated test for it.