net: free socket memory in per-socket cleanup fns
With TCP sockets we need to maintain state after the socket is cleaned
up kernel-side in some cases (e.g. to hold the socket's address while in
TIME_WAIT).  Rather than create separate state, this allows us to just
use the socket state itself even after the socket API part of the kernel
is done with it.