user: add getsockopt/setsockopt syscalls
The general case here is interesting --- the buffer points to an
arbitrary type that depends on the specific socket option, so the DMZ
code cannot trivially convert between 32 and 64-bit versions (or a more
generic user-ABI-dependent conversion, if that were ever a thing).
Since most socket options are ints, for now just pass it through.  For
timeval socket options we can do a check in the socket option handlers
for the 32-bit or 64-bit version.

A generic implementation of this might either (a) add a helper that the
wrapper can call beforehand to sniff the intended type to do the
conversion (e.g. sockopt_type(..)), or (b) even more generically pass
the buffer to the sockopt method as some new userbuf_t structure that
encodes the user buffer address, length, AND userspace ABI; the syscall
implementation could then call an appropriate converter on the userbuf_t
to copy and convert to/from the appropriate kernel-space ABI depending
on the type.  Definitely overkill for the time being.