tcp: add basic TCP socket lock structure
This is preemptive --- I figure it is easier to start building in the
required concurrency/locking now rather than needing to add later.

Each TCP socket gets two locks, a mutex and a spinlock.  The mutex
protects the overall socket state (and also will be used for blocking
operations along with the wait queue).

The spinlock is used for any state that is touched by a deferred
interrupt (i.e. the incoming packet handling code).