proc: add spinlock variant of scheduler_wait_on_locked
I have somewhat mixed feelings about this --- it feels somewhat strange
to want to yield when holding a spinlock.  On the other hand, this is
the only correct way (with SMP) to implement a condition variable that
can be triggered by a deferred interrupt --- since the defint could run
either on the current core (and must be blocked in the critical section
to prevent deadlock, by disabling interrupts) or on a different core
(and must be blocked to prevent race conditions, with mutual exclusion
--- i.e. the spinlock itself).

The pattern for use will be:
 1) for code that does not interact with defints or interrupts, continue
    to use kmutex.
 2) for code that interacts with defints (and today would use
    DEFINT_PUSH_AND_DISABLE() etc), switch to a spinlock and
    scheduler_wait_on_splocked().
 3) optionally, use _both_ a mutex and a spinlock --- the mutex protects
    around heavyweight operations between threads, and the spinlock
    protects a small amount of defint-safe state.

For (3), we could probably have a scheduler_wait_on() variant that
atomically locks/unlocks _both_ a mutex and a spinlock.  I'm not
convinced that's necessary or worth the complexity.

(2) is the meat of this change --- it replaces the current global
defint-based locking with more a granular defint signal (not that that
is implemented today), AND makes it SMP-safe by protecting against
concurrent access by other cores (running either normal code or defint
code).