tcp: implement basic connect/FIN/close handling
This implements a very small subset of the TCP state machine and adds a
basic test to traverse it.  The primary value to this commit is laying
the groundwork for the rest of the state machine and socket lifecycle
management (which is complex).

Sockets are now refcounted, with VFS, each TCP data structure, and
defints holding refcounts each.  This significantly simplifies lifetime
management as it is possible for any those to retain the last reference
to a socket in the end state.