tcp: consider FIN part of segment for window check
This corrects the behavior where the FIN would not be considered as part
of the received segment --- if a packet came in with retransmitted data
and a FIN matching recv_next, the FIN would be ignored.  RFC 9293
specifies that FIN (and SYN) should be considered part of the segment
length when it comes to window calculations (even though they don't
occupy buffer space), so this now conforms to that.