net: fix memory leak when ip_send() fails
Currently the lower levels of the stack only consume the packet on
success.  All callers of ip_send() currently assume that it
unconditionally takes ownership of the pbuf_t, so this makes that
the case for simplicitly.

You could imagine a protocol stack such as TCP saving the pbuf_t for a
retransmission on failure (rather than reconstructing), but it doesn't
currently do that and it's not clear it would be worth the complexity.

This is currently only exercised by the new TCP TAP-based test, which
causes the ARP system to return -EINTR.