tcp: implement SO_REUSEADDR in sockmap (not socket, yet)
This means we do a linear search across all sockets bound to a
particular address and port every time bind() is called with a 3-tuple.
This could be avoided with a more complicated data structure that
explicitly kept reusable and non-reusable bindings separate (and
therefore could simply be checked for presence).  A counter would also
work.

For the any-addr vs not-any-addr matching, a different approach to avoid
the linear searches would be needed.  For example, only non-any-addr
bindings could be kept in the port map (any-addr:port can be trivially
looked up), but that would complicate other code.