tcp: close socket when unable to connect due to address conflict
For some reason, this causes tests to take an extra second to run,
somewhere in net_connect().  If I enable logging, or change various
other parts of the code, the delay goes away.  I think this is the same
qemu delay issue I had several years ago that I never understood.

Ignoring it for now, hopefully it goes away as I edit the code more.