tcp: use incrementing ephemeral ports
This is the explanation for the 1-second increase in test run time in
commit f59c69dce97806c39c836 --- there is a race condition where the
next socket created binds to the same local address as before, connects
to the server address, which goes to the server socket in TIME_WAIT,
which sends a challenge ACK, which triggers a RST, and then we need to
wait for the SYN to retransmit (1s later) to complete the connection.

So, the current behavior is correct protocol-wise, but using an
incrementing ephemeral port minimizes the likelihood of rapid 5-tuple
reuse.