tcp: handle race when socket closes before dispatch action
For the most part the code already does the right thing, but this fixes
a couple of cases.

This structure with actions isn't great --- I think for most of these
cases, there is only ever one action taken.  And for most actions, we
promptly re-lock the socket's spinlock, which is what allows these races
to happen.

I should reexamine this structure and try and make it so that actions
are started with the lock still held from the first part of the
dispatch.  This will mean that the action helpers (such as
tcp_send_datafin) will have to take the socket with the lock held and
unlock it, which is a strange contract --- but it's already the case for
some helpers, and I think it's a weak reason to introduce the race
conditions inherent in the unlock-the-immediately-relock sequence.