tcp: refactor packet handling with terminal actions
Currently the code is structured assuming that multiple actions may be
taken at the end of processing an incoming packet.  In reality, with
only one exception (TCP_DROP_BAD_PKT + TCP_SEND_ACK), they are all
mutually exclusive.

This simplifies the control flow by changing the action bitmask to be a
single terminal action, with an optional bit to trigger sending an ACK.
This simplifies the code, and will make it much easier to restructure
the actions such that we don't need to unlock and relock the socket
mutex (and deal with the associated race conditions).