net/rtl8139: cap recv packets per defint to prevent tx starvation
This is a pretty crude approach, but works OK.  The current problem is
that when there are many packets pending, we will recv() them all in
a loop --- if the RX path causes transmission (ACKs, for example), then
our TX descriptors will fill up and we won't have an opportunity to deal
with them until after we've processed all the received packets.  This
results in losing ACKs.

Easily tested by transferring a moderately-sized filed between guest and
another endpoint.

There are other issues here that should be solved, ideally.  For one,
the defints can queue up unbounded if the RX path takes too long ---
this should be solved more generally with a defint queuing change, the
UART code already deals with this as well.  Additionally, we could
simply check the TX descriptor status when necessary rather than waiting
for the TX IRQ to fire; since we don't queue packets in the NIC driver,
we actually don't care about when the TX descriptor frees up and could
just do it directly on the TX path.

With this change, transfers of large files over TCP still has issues due
to silly window syndrome, but at least we don't drop ACKs.

This also includes a flag printf spec for the TX status register, which
is probably not generally that useful but no reason to make myself write
it again in the future.