tcp tests: define different address for loopback tests
With raw-socket based testing, we can use the same addresses for the
direct-packet-munging tests and the normal loopback TCP-to-TCP-stack
tests --- if a TCP socket is connected, it will get the packets before
the raw sockets ever see it.

With TUN-based testing, however, packets sent to the TUN device always
disappear into a black hole if the test isn't listening (this is overall
a good thing).  So separate out tests that need to talk to themselves
over the loopback address and give them dedicated address constants.