tcp: don't send FIN if window is closed
This is slightly ambiguous per the RFC --- RFC 9293 says, on the one
hand, that the window is "The number of data octets beginning with the
one indicated in the acknowledgment field that the sender of this
segment is willing to accept."  This implies that the window does _not_
include control bits like SYN and FIN (and therefore a FIN should be
transmitted if the window is zero).  This makes sense --- the window is
for controlling buffer space, and a FIN doesn't take any buffer space.

On the other hand, the RFC also says "Note that when the receive window
is zero no segments should be acceptable except ACK segments." and
defines "receive window" as "This represents the sequence numbers the
local (receiving) TCP endpoint is willing to receive. Thus, the local
TCP endpoint considers that segments overlapping the range RCV.NXT to
RCV.NXT + RCV.WND - 1 carry acceptable data or control. Segments
containing sequence numbers entirely outside this range are considered
duplicates or injection attacks and discarded."

Linux does not transmit a FIN when the window is closed.  macos does,
however this does not interact well with a receiving Linux host (which
sends an ACK that doesn't acknowledge the FIN, causing macos to
retransmit, and this goes in a cycle).

I'm choosing the more conservative behavior, consistent with Linux.