tcp: deflake TIME_WAITs in real socket tests
These currently flake for two reasons:
1) sockets take too long to leave TIME_WAIT, causing them to spuriously
   conflict with later tests.
2) sockets leave TIME_WAIT too _soon_, racing with them shutting down.
   This causes them to fail to send a final ACK, leaving their peer
   socket stuck in LAST_ACK indefinitely (they will never leave).  This
   interferes with ~everything.

(1) can be solved with more generous sleeps, but (2) cannot.

Instead, this adds a deterministic way (without sending a RST, which is
hard in a real-socket-pair test) to force a socket in TIME_WAIT to
close, but calling setsockopt(SO_TCP_TIME_WAIT_LEN, 0).  The real socket
tests are updated to use this and now are not flaky.

There are other tests that set SO_TCP_TIME_WAIT_LEN to small values ---
I left a TODO to audit them and see if they should be updated to use
this method as well.