tcp: fix timeout overflow race condition
This is awkward, but matches how this is handled in other parts of the
codebase.  The bug is that in rare occasions the clock will increment in
between when |now| is calculated and when the (absolute) end timeout is
calculate inside scheduler_wait_on_splocked().  In that case the end
time will overflow and become small, and the timeout will trigger
immediately.

It might be better to have the scheduler_wait...() calls take an
absolute timeout rather than a relative one; in a quick survey I think
most call sites are passing relative deadlines calculated from absolute
ones.