tcp: remove unneeded set(SO_TCP_TIME_WAIT_LEN) in tests
Most of these tests don't have the socket go into TIME_WAIT, or if they
do, they send a RST to force it to close.  Therefore setting the
TIME_WAIT timeout is at best extraneous, and risks flaking the test.

Also remove a stale TODO, and the TODO to reexamine all of the tests
that set SO_TCP_TIME_WAIT_LEN.