tcp: fix receive-before-send-SYN race
There are several elements to this fix:
1) hold a ref across tcp_send_syn() in tcp_handle_in_listen() --- this
   avoids a use-after-free if the socket is closed (removed from the
   socket map) before tcp_send_syn runs.
2) handle any state in tcp_send_syn() --- we cannot count on the socket
   not having transitioned states in between the caller unlocking the
   mutex and us locking it.
3) don't attempt to tcp_send_datafin() if not in a TCPSTATE_ESTABLISHED
   state.  This can happen if we get a packet while the socket is in
   SYN_RCVD (and currently triggers an assertion failure at
   protocol.c:156 when we try to calculate the amount of unacked data).
   If not in a TCPSTATE_ESTABLISHED state, no point in trying to send
   data anyway.
4) fix bug in handling an ACK in TCP_SYN_RCVD.  This currently simply
   checks to see if there is no unacked data --- but in this race
   condition, there may be no unacked data (if the SYN isn't sent), but
   the SYN isn't yet ACK'd.  Now that we track syn_acked explicitly, use
   it here.

A better solution to all of this complexity would be to simply have
tcp_send_syn() take a locked socket, and avoid the race condition
altogether.  However due to the lock ordering in the LISTEN flow, and
the fact that we can't unlock spinlocks in a different order than we
locked them (a limitation in the spinlock impl), we have to unlock and
relock the child socket, at least currently.  A refactor of that flow
would eliminate this race condition and complex tests.