vfs: use fd mode for poll() on FIFOs/pipes
This matches the behavior on Linux for both FIFOs (named) and anonymous
pipes --- the read end of the pipe (or a FIFO opened with O_RDONLY) will
never return POLLOUT (etc), and vice versa.

On macos, FIFOs (named pipes) behave like normal files (not respecting
fd mode), while (anonymous) pipes behave the same as here.  That doesn't
seem to be POSIX-compliant, but it's not 100% clear.