vfs: fix RLIMIT_NOFILE bug in dup2()
dup2() currently incorrectly casts the RLIMIT_NOFILE value to an int (from
rlim_t, which is unsigned long).  If the limit value is larger than INT_MAX but
not equal to RLIM_INFINITY, it will overflow (potentially to a negative number)
and the comparison with the target fd will not function correctly.

This also manifests when a 32-bit userspace program sets the limit to
RLIM_INFINITY, which is not currently handled correctly by the 32-to-64
syscall wrappers --- this is a separate bug.  Due to this, the 32-bit
value of RLIM_INFINITY is passed through unmodified, and therefore does
not hit the check if the limit is equal to RLIM_INFINITY --- which then
triggers the overflow-int bug fixed here.