test: force-flush block cache at end of test runs
This currently manifests as repeated failures when running test_all.sh with
multiple runs due to the existence of a _tun_test_dev file.  I believe this is
due to the file existing still existing on-disk at the end of the test run
(because the unlink() in tcp_test.c hasn't flushed to disk yet).  When run
manually, I can't repro this; adding the force-flush and small sleep at the end
of the overall test run fixes it, though.