net: fix TUN/TAP dev id generation
It needs to pass DEVICE_ID_UNKNOWN in order to generate a new ID; currently, it
passes zero, which means that only one TUN/TAP device can be created at once.

Also, fix a memory leak in the NIC creation code (if a name collision is
detected, the reference is not properly released.