net: pass sockaddr_storage to getsockname/getpeername
This was not a good API choice in the first place --- at least a couple of times
I've screwed up and passed something smaller than a sockaddr_storage to these,
causing stack corruption.

The alternative is to pass a socklen_t as well (per the userspace API) --- I
decided against that because it's clumsy (nearly all existing call sites pass a
sockaddr_storage already!), and requires more complicated logic in each socket
implementation.

This approach keeps the contract the same, but updates the type signature to
match (and enforce).