net: trim extra bytes from IPv6 packets before passing to protocol
Discovered this bug when testing NDP with qemu, which (for some reason) causes
extra bytes to show up at the end of the received ethernet frame --- this causes
the ICMPv6 checksum validation to fail.

For now, pbuf_trim_end() is unidirectional --- it could be made reversible by
tracking the buffer size independent of the data size (as with the start
padding).