net: check dest sa_family before checking size in UDP sendto()
This lets us upgrade the check in sizeof_sockaddr() to a DFATAL if an invalid
address family is passed.  Also, update a UDP test that I think was incorrectly
written in the original IPv6 CL, as its the exact same as the statement just
before it (and there's no reason to test each of these twice).  Changed it to
check for a "too-short" address length with a valid address.

Also update the check in the UDP code to assert that the address family can be
read, as the length should be checked in net_sendto() (and therefore this branch
is untestable).