net: ignore IPv6 flow and scope in sockaddr_equal()
Flow label should definitely be ignored.  I'm slightly undecided on
sin6_scope_id.  I played around with having the scope ID be considered here, but
it's not clear to me that that is semantically correct.  It seems like the scope
ID should only be considered on outbound connections as a "hint" to interface
selection.

On the one hand --- if I ignore it now, it will be hard to make it required
later.  On the other hand, I'm not sure how much it should matter.  AFAICT, for
Linux, it's only used for link-local addresses and otherwise ignored.

The generalized form of the scope ID is, I think, a VRF.  So I could implement
VRF support (similar to Linux) in the future.