rtl8139: calculate multicast filter rather than accepting all
Also, switch to writing the filter in two 32-bit writes rather than eight 8-bit
writes.  The data sheet indicates this is how it must be done, though qemu
doesn't care.

This is not exercised normally with outbound IPv6 connections (the easiest way
to test this with qemu).  To test this, I statically inserted a neighbor cache
entry for fe80::2 (the default IPv6 gateway currently), so that to send a reply
SYN/ACK qemu needs to send an NDP neighbor solicitation, exercising the
multicast filter.