devicetree: handle long strings properly
The _real_ bug was fixed in commits fcdfbbd8 and d0e82475, which resolved a
buffer overflow in this code that existed previously.

This manifested when upgrading qemu to version 9.0.2 --- somewhere along the way
qemu started using longer strings for the riscv,isa property of the CPU.  This
in fact was already causing a buffer overflow, but it seemed to (somehow) be
harmless.  When the string was lengthened to over 200 characters, it stomps the
stack and causes a crash on startup on riscv64.

I worked around this for a while with the `-cpu rv64,short-isa-string=on' flag
to qemu (which is for exactly this problem), but this (and the ksnprintf
commits) fix it properly.