common: check logging enabled before formatting log line
This avoids doing all the work of ksnprintf() if we're just going to throw away
the string anyway.  During a `test all` run, without this, klogfm() accounts for
over 90% of time spent in kvsnprintf().  After this change, it's <2%.

On a test run, kvsprintf() before this is about 1.5% of total test time
(~450ms).  Most of that, per above, is klogfm() calls that ultimately are thrown
away (62% are from int_handler()).  After that it's a long tail of calls that
are also not usually printed (syscall handling, TCP stack, block cache, etc).

This is not a significant optimization, but might as well (especially since klog
is ~2x more expensive now, with snprintf bounds checking).