hashtable: use 64-bit keys
When hashing stack traces, I'm frequently seeing collisions --- the birthday
problem estimates that with a 32-bit key space, we'll have a 50% chance of
collision around 2^16 entries (65k).  I'm seeing collisions generally between
20k and 60k entries in the stack trace tables, which is consistent with that.

By using 64-bit hashes, we should be able to store far more stack traces without
chance of key collision than before, which is required for profiling to be
effective.

On i586, there is no 64-bit division operation --- it requires a non-existent
compiler builtin.  To calculate the hash table bucket, we truncate the 64-bit
key to 32 bits before mod'ing with the bucket count.  If the bucket count is a
power of 2 (less than 32 bits) this yields the same result.  It's possible
there's a number-theoretic reason this is problematic for non-powers-of-two, but
I can't think of it.