profiler: switch tracetbl to custom hash table
The tracetbl is not, as it turns out, a good fit for CPU profiling.  Its static
size is too small by default --- in one test run profiling the entire test suite
the profiler needed to store 170k unique stack traces.  Additionally, the
refcount grows much beyond 2^15 (short max) for frequent stack traces (such as
the idle thread).

This requires a cheap allocator like an arena allocator (though we never free it
in this case) --- we end up doing tens of thousands of small allocations (which
we never free), which algorithm destroys the O(n^2) performance of the default
kmalloc algorithm; it's impossible to profile long-running test runs because the
whole program grinds to a halt.

Because kmalloc is interrupt-safe, it is safe to use the standard arena
allocator (which must periodically allocate new blocks).  It would be possible
to make this entirely bypass kmalloc with a static buffer allocator of some sort
if necessary.