scheduler: add idle time profiler
During a normal test run, a significant time (possibly over 80%) is spent in the
idle thread, per normal CPU profiling.

This adds a PROFILE_IDLE mode that measures how long is spent idling, and
records the stack trace of the thread that is first scheduled when the idle
finishes.  The assumption is that it is this thread that triggered the idling
(e.g. by sleeping, or waiting on an interrupt-triggered event).

Unfortunately there's no obvious low-hanging fruit revealed by this, but it's
interesting and useful nonetheless.