linker: place .lbss (and other) sections before KERNEL_END_SYMBOL
This ensures .lbss, .lrodata, and .ldata are properly placed before
KERNEL_END_SYMBOL.  Previously, none of these would be generated.  When updating
to GCC 14.2.0, however, the g_tracetbl symbol (a large array) is placed in
.lbss.  This is caused by GCC commit 1a64156c7e25813afa8d4d8bbeb0590cad91cf55.

The linker places .lbss at the end of the linked object file, but
KERNEL_END_SYMBOL is before it.  That means when we allocate pages during
paging_init()/mem_init(), we select pages that overlap with g_tracetbl to place
the page tables in.  This causes x86_64 to crash on boot the first time
g_tracetbl is used.

.ldata and .lrodata are not currently generated, but including them for
consistency and future-proofing.