net: deflake echo test with read-delay loop
Rather than having a static delay before trying to read the echoed data, try in
a loop.  This allows the test to pass when there is a system delay causing the
echo to take longer than 20ms, without unnecessarily extending the test
duration.

Flakes observed in test round sponge-20241108-185716, though not easily
reproducible.