tcp: don't send window updates in TIME_WAIT
Currently if enough data is read to generate a window update, one is sent even
if the socket is in TIME_WAIT.  If it's in TIME_WAIT, however, the other side
isn't listening anymore, and a RST will likely be generated.

Caught when writing a multi-threaded echo test, where the reader thread reads
until the writer closes the connection --- this triggered the TIME_WAIT window
update scenario.