proc: add preemption hooks in core data structures
The inspiration for this is commit fac786865b1 --- to induce this bug, you need
to hit a preemption at _just_ the right moment.  Rather than add whitebox hooks
to all code to try and trigger these (which would never work), this adds
optional preemption hooks to two core data structures, list and hashtable.  When
enabled, preemptions will be induced inside those data structures.  If access to
the data is not properly protected, this significantly increases the likelihood
of observing a race and triggering an assertion failure.

This consistently triggers the block cache bug fixed in fac786865b1.  I tried
also inducing a race condition in the TCP code (harder to find), and it triggers
there as well.  It also triggers on a race condition I tested adding in
block_cache_get().

This is still way less effective than a proper sanitizer, but makes the
multithreaded tests for preemption-safe code much more effective.