common: make refcount_t interrupt-safe
refcount_t is used by the TCP code to refcount sockets --- including across
interrupt/defint handlers.  So this must be interrupt-safe.

This incurs a cost to other refcount users, but not enough to worry about at
this point.  And this should really be an atomic anyway.

Caught with static path analysis with the call graph analyzer! \o/