tcp: implement zero-window probes
This also surfaced during the multithreaded test writing.  On occasion, the very
last set of window updates would be a zero-window followed by an open-window,
and due to thread scheduling the packets would get reordered.  This results in
the zero window being the last window update received, and the connection
deadlocking.