vfs: interrupt-protect file and FD table updates in vfs_open_vnode()
This is a cheap hack to make these interrupt-safe.  Overall the process and VFS
data structures (other than the vnode table?) are _not_ preemptable.  However
this is one specific area that occasionally causes problems when running the TCP
multithread tests with preemption enabled, when there are races on FD opening.

This should be fixed more broadly, but for now just make these specific sections
interrupt-safe to deflake those tests.