riscv64: add v1 of double fault (stack overflow) handler
This version is reasonably clear but doesn't work properly with stack unwinders,
and is probably a bit less efficient than it could be.

To unwind the stack, we'll need to store the original sepc (interrupted address)
in the thread's scratch space.  Then either extract it explicitly in the double
fault handler (or unwinder), or set up a fake stack frame on the double fault
stack.