riscv64: add v2 of double fault handling
This fixes a stack issue in the previous version (was not correctly starting at
the top of the double-fault stack --- so was trampling arbitrary memory).  It
also sets up a couple fake stack frames for the unwinder, and adds support to
the unwinder to detect the dauble-fault stack and allow unwinding across it.