tcp: only retransmit after exactly 3 duplicate ACKs
I believe this is RFC-compliant --- further duplicate ACKs should affect the
cwnd, but not cause additional fast retransmits.  If the fast retransmit is
lost, it must wait for the retransmit timer to trigger.

This avoids a pathological case when there's a closed receive window that stays
closed for a while --- the sender sends a zero-window probe, which gets back a
dup-ack.  If that happens three times in a row, a fast retransmit of the ZWP is
triggered, which then causes a doom spiral of ZWPs/dupacks/fast retransmits.

For some reason, this only manifests in the multithreaded tests when running
with logging enabled (maybe logging slows it down enough to cause the fast
retransmit?).  There is another pathology with the loopback defint scheduling
here --- each ZWP/dupack triggers generation of a response packet, which keeps
the loopback queue full.  Eventually the loop returns (after 50 exchanges), but
schedules another defint --- which immediately runs again.  This starves all
other work, including the thread(s) that would read from the socket and open the
window up again.