net: add throttling to loopback packet processing
This adds two throttling mechanisms to the loopback code:
1) defints are only scheduled if there is not currently a defint processing
   packets.  This prevents a doubling of defint scheduling, where a situation
   that causes rapid 1-for-1 packet generation will cause both the call to
   loopback_send() as well as the backoff at the end of the defint loop to
   schedule a followup defint; this causes defint overflow.

2) throttling on the defint processing itself.  When the maximum number of
   packets are processed, the loopback won't process additional packets for a
   cooldown period, allowing other threads (which may remove the source of the
   packet jam, or otherwise back off) to run.

Together, these avoid the kernel panic caused by the previously-fixed TCP
retransmission of ZWPs issue (i.e. they successfully throttle the creation of
defints and processing of packets such that the death spiral --- and panic due
to defint overflow --- don't occur).

(1) would not be an issue if we moved to a defint system that only scheduled one
defint per type at once.  A lot of modules have similar logic, so we should
probably do that --- there's no need for arbitrary defint queueing.