tcp: bypass TUN FD in tests; write directly to chardev
The VFS path requires doing an FD lookup, which will (soon) require locking a
mutex.  That means that if we inject packets during a defint, we could block,
which is not allowed.

This generally should not be possible (defints should not be manipulating file
descriptors), but in the tests we inject packets during a test_point_add() hook,
which can run during a defint.  In this specific case, blocking is likely
harmless due to the test construction, but I don't think it's generally safe,
and may not be safe in the future depending on what other threads are running.