tcp: don't write() from ZWP race test hook
This is unfortunate, but the test as written is currently illegal (once the FD
lookup code is properly protected with a lock).  It also tests a race that can't
happen, as without SMP, this would require preemption to happen while a defint
is running, which is disallowed (currently).

Once SMP or defint-preemption is implemented, this race will be possible, and
there will be a mechanism to trigger it from a test hook.