vfs: add g_file_table mutex
Because the file entries are refcounted, and the table is only mutated when a
file is created or deleted, we only need to use this mutex when searching for an
empty spot, or allocating/deallocating a slot.  If we are looking up a file
descriptor, we don't need to as the entry is guaranteed to be static.