linker: place .srodata/.sdata/.sbss properly before KERNEL_END_SYMBOL
This has always been a bug, but didn't start to manifest until I upgraded my
clang to version 19.1.0.  Clang commit b7f97d3661 started placing small data in
a .srodata.cst8 section --- though this is not used, it caused the g_queue_len
symbol (in .sbss) to be placed after KERNEL_END_SYMBOL, which in turn caused it
to overlap with DMA mappable page ranges.

At the current HEAD, this causes a kernel panic when the RTL NIC driver sends a
particular packet, causing the NIC to overwrite the memory containing
g_queue_len (which is in its rxbuf).  This causes a panic the next time a defint
is scheduled.