riscv: fix user-mode linking with clang
clang cannot link riscv64 user-mode binaries --- by default, it fails because it
cannot find ld.lld (the default linker).  If you override the linker with
-fuse-ld to point to ld or gcc, it still fails (due to missing compiler-rt
builtins and/or flag mismatches).

This overrides the linker to always be the target-specific GCC.  On non-riscv64
targets (i586 and x86_64), this is already the case, but it goes via clang
(clang uses gcc to link).  This forces the linker to always be gcc, which is
invoked directly (rather than via clang).

LD= was the wrong variable to use this whole time --- LD is not an scons
variable.  LINK= is the correct variable.

This also requires setting -march=rv64gc to enable hardware floating point, to
allowing linking between the user code and newlib.