nvme: use spinlock to avoid defint race
DEFINT_PUSH_AND_DISABLE()/DEFINT_POP() is not safe to use across a critical
section when preemption is enabled.  The thread may be preempted before calling
scheduler_wait_on_interruptable() --- the next thread may have defints enabled,
run a defint, and then we miss the wakeup and sleep until we time out.

This is not (for NVMe) a correctness issue, as the transactions are actually
completed.  See example run sponge-20241213-223637 (run user_tests with NVME
logging set to 3, and the ktest quiesce timeout set to 20000ms).  In 13/100
runs, the flush queue thread gets preempted triggering this race.  This causes
the block to be marked not flushed (as the write() operation returns ETIMEDOUT),
and significantly extends the quiesce period (enough to cause a timeout with
normal parameters).

Rather than update DEFINT_PUSH_AND_DISABLE() to disable preemption, I'm just
going to replace all uses of it with proper spinlocks --- this is what is needed
anyway to prepare for SMP.

I'm not totally sure why I didn't use the spinlock here the whole time; perhaps
this code predates scheduler_wait_on_splocked()?