proc: fix scheduler timeout race condition
Fun!  This bug dates back at least until 2015 (commit e1d206e92f2a832baaae3978d,
which introduced timeouts to waits), but possibly back to the origin of the
scheduler code in 2013, if any modules called scheduler_wake_*() from an
interrupt context (which is likely).

This was observed in sponge-20241214-164013-riscv64-clang-0018, with the
following stack trace:
Thread 2 (crashing):
 #0 0xffffffff8020ab02 in die() [kassert.c:57]
 #1 0xffffffff8020ac30 in kassert_msg() [kassert.c:72]
 #2 0xffffffff80233024 in scheduler_timeout() [scheduler.c:125]
 #3 0xffffffff80478622 in internal_timer_handler() [timer.c:105]
 #4 0xffffffff80474890 in rsv_timer_interrupt() [timer.c:73]
 #5 0xffffffff80471fe8 in int_handler() [interrupts.c:121]
 #6 0xffffffff8020a1a0 in int_handler_asm() [interrupts_asm.s:220]
 #7 0xffffffff80231d04 in kthread_queue_pop() [kthread.c:347]
 #8 0xffffffff80232db4 in scheduler_wake_all() [scheduler.c:218]
 #9 0xffffffff8048342e in txn_done() [controller.c:322]
 #10 0xffffffff80484830 in nvmec_check_queue() [controller.c:202]
 #11 0xffffffff804845c2 in nvmec_defint() [controller.c:215]
 #12 0xffffffff8023034e in defint_process_queued() [defint.c:89]
 #13 0xffffffff8023022c in defint_set_state() [defint.c:61]
 #14 0xffffffff80234006 in kspin_unlock() [spinlock.c:64]
 #15 0xffffffff80232bb4 in scheduler_wait_on_internal() [scheduler.c:166]
 #16 0xffffffff80232cf6 in scheduler_wait_on_splocked() [scheduler.c:202]
 #17 0xffffffff8048330e in nvme_submit_blocking() [controller.c:342]
 #18 0xffffffff80527488 in nvme_bd_op() [block_dev.c:77]
 #19 0xffffffff805271f2 in nvme_bd_write() [block_dev.c:97]
 #20 0xffffffff80485b0a in bd_write_page() [memobj_block_dev.c:108]
 #21 0xffffffff80216bd6 in flush_cache_entry() [block_cache.c:183]
 #22 0xffffffff8021714e in flush_queue_thread() [block_cache.c:245]
 #23 0xffffffff80231890 in kthread_trampoline() [kthread.c:82]
 #24 0xffffffff804737d8 in riscv_kthread_trampoline() [kthread_asm.s:19]
PANIC: assertion failed: kthread_current_thread()->queue == 0x0 (proc/scheduler.c:125)

The bug is simply a race on the kthread_queue_t --- it is popped off, but the
interrupt fires before the thread is actually made runnable.  This can result in
an assertion failure (as above), or data structure corruption.

This adds two kthread tests --- they catch the bug, but only if
CONCURRENCY_TEST_ITERS_MULT is cranked up to 100 or so, unfortunately.